+++ example/integration_test/find_interaction_controller/main.dart
@@ -4,7 +4,8 @@ import 'package:flutter_test/flutter_test.dart';
 import 'find_interactions.dart';
 
 void main() {
-  final shouldSkip = kIsWeb;
+  final shouldSkip =
+      kIsWeb || [TargetPlatform.macOS].contains(defaultTargetPlatform);
 
   group('FindInteractionController', () {
     findInteractions();