+++ example/integration_test/in_app_browser/open_file_and_close.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/foundation.dart';
-import 'package:flutter/services.dart';
 import 'package:flutter_inappwebview/flutter_inappwebview.dart';
 import 'package:flutter_test/flutter_test.dart';
 
@@ -34,12 +33,12 @@ void openFileAndClose() {
     await inAppBrowser.firstPageLoaded.future;
     var controller = inAppBrowser.webViewController;
 
-    final String? url = (await controller.getUrl())?.toString();
+    expect(controller, isNotNull);
+    final String? url = (await controller!.getUrl())?.toString();
     expect(url, endsWith("in_app_webview_initial_file_test.html"));
 
     await inAppBrowser.close();
     expect(inAppBrowser.isOpened(), false);
-    expect(() async => await inAppBrowser.webViewController.getUrl(),
-        throwsA(isInstanceOf<MissingPluginException>()));
+    expect(inAppBrowser.webViewController, isNull);
   }, skip: shouldSkip);
 }