+++ example/integration_test/in_app_webview/take_screenshot.dart
@@ -1,7 +1,6 @@
 import 'dart:async';
 
 import 'package:flutter/foundation.dart';
-import 'package:flutter/widgets.dart';
 import 'package:flutter_inappwebview/flutter_inappwebview.dart';
 import 'package:flutter_test/flutter_test.dart';
 
@@ -21,27 +20,23 @@ void takeScreenshot() {
         Completer<InAppWebViewController>();
     final Completer<void> pageLoaded = Completer<void>();
 
-    await tester.pumpWidget(
-      Directionality(
-        textDirection: TextDirection.ltr,
-        child: InAppWebView(
-          key: GlobalKey(),
-          initialUrlRequest: URLRequest(url: TEST_CROSS_PLATFORM_URL_1),
-          onWebViewCreated: (controller) {
-            controllerCompleter.complete(controller);
-          },
-          onLoadStop: (controller, url) {
-            pageLoaded.complete();
-          },
-        ),
-      ),
+    var headlessWebView = new HeadlessInAppWebView(
+      initialUrlRequest: URLRequest(url: TEST_CROSS_PLATFORM_URL_1),
+      onWebViewCreated: (controller) {
+        controllerCompleter.complete(controller);
+      },
     );
+    headlessWebView.onLoadStop = (controller, url) async {
+      pageLoaded.complete();
+    };
+
+    await headlessWebView.run();
+    expect(headlessWebView.isRunning(), true);
 
     final InAppWebViewController controller = await controllerCompleter.future;
     await pageLoaded.future;
 
     await Future.delayed(Duration(seconds: 1));
-    await tester.pump();
 
     var screenshotConfiguration = ScreenshotConfiguration(
         compressFormat: CompressFormat.JPEG,