+++ example/integration_test/webview_flutter_test.dart
@@ -33,13 +33,13 @@ void main() {
   FindInteractionController.debugLoggingSettings.usePrint = true;
   FindInteractionController.debugLoggingSettings.maxLogMessageLength = 7000;
 
-  in_app_webview_tests.main();
-  find_interaction_controller_tests.main();
+  // in_app_webview_tests.main();
+  // find_interaction_controller_tests.main();
   service_worker_controller_tests.main();
   proxy_controller_tests.main();
   headless_in_app_webview_tests.main();
   cookie_manager_tests.main();
   in_app_browser_tests.main();
-  chrome_safari_browser_tests.main();
+  // chrome_safari_browser_tests.main();
   in_app_localhost_server_tests.main();
 }